test: Fix tests for split up manage provider
