+++ packages/app_center/lib/manage/manage_page.dart
@@ -3,13 +3,10 @@ import 'package:app_center/error/error.dart';
 import 'package:app_center/l10n.dart';
 import 'package:app_center/layout.dart';
 import 'package:app_center/manage/local_snap_providers.dart';
-import 'package:app_center/manage/manage_l10n.dart';
+import 'package:app_center/manage/manage_snap_tile.dart';
 import 'package:app_center/snapd/snapd.dart';
-import 'package:app_center/store/store.dart';
-import 'package:app_center/widgets/widgets.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:snapd/snapd.dart';
 import 'package:ubuntu_widgets/ubuntu_widgets.dart';
 import 'package:yaru/yaru.dart';
 
@@ -23,8 +20,19 @@ class ManagePage extends ConsumerWidget {
   @override
   Widget build(BuildContext context, WidgetRef ref) {
     final updatesModel = ref.watch(updatesModelProvider);
+    final filteredLocalSnaps = ref.watch(filteredLocalSnapsProvider);
     final l10n = AppLocalizations.of(context);
     final textTheme = Theme.of(context).textTheme;
+
+    if (updatesModel.hasError || filteredLocalSnaps.hasError) {
+      return ErrorView(
+        error: updatesModel.error ?? filteredLocalSnaps.error,
+        onRetry: () {
+          ref.invalidate(updatesModelProvider);
+          ref.invalidate(filteredLocalSnapsProvider);
+        },
+      );
+    }
 
     return Padding(
       padding: const EdgeInsets.symmetric(vertical: kPagePadding),
@@ -83,7 +91,7 @@ class ManagePage extends ConsumerWidget {
                       ),
                   ];
                 },
-                error: (error, stack) => [ErrorView(error: error)],
+                error: (_, __) => [],
                 loading: () =>
                     [const Center(child: YaruCircularProgressIndicator())],
               ),
@@ -92,18 +100,17 @@ class ManagePage extends ConsumerWidget {
           updatesModel.when(
             data: (refreshableSnaps) => SliverList.builder(
               itemCount: refreshableSnaps.length,
-              itemBuilder: (context, index) => _ManageSnapTile(
+              itemBuilder: (context, index) => ManageSnapTile(
                 snap: refreshableSnaps.elementAt(index),
-                position: determinePosition(
+                position: determineTilePosition(
                   index: index,
                   length: refreshableSnaps.length,
                 ),
                 showUpdateButton: true,
               ),
             ),
-            error: (error, stack) => SliverToBoxAdapter(
-              child: ErrorView(error: error),
-            ),
+            error: (error, stack) =>
+                const SliverToBoxAdapter(child: SizedBox.shrink()),
             loading: () => const SliverToBoxAdapter(
               child: Center(child: YaruCircularProgressIndicator()),
             ),
@@ -181,28 +188,22 @@ class ManagePage extends ConsumerWidget {
               const SizedBox(height: 24),
             ],
           ),
-          Consumer(
-            builder: (context, ref, _) {
-              final filteredLocalSnaps = ref.watch(filteredLocalSnapsProvider);
-              return filteredLocalSnaps.when(
-                data: (data) => SliverList.builder(
-                  itemCount: data.length,
-                  itemBuilder: (context, index) => _ManageSnapTile(
-                    snap: data.elementAt(index),
-                    position:
-                        determinePosition(index: index, length: data.length),
-                  ),
-                ),
-                error: (error, stack) => SliverToBoxAdapter(
-                  child: ErrorWidget(error),
-                ),
-                loading: () => const SliverToBoxAdapter(
-                  child: Center(
-                    child: YaruCircularProgressIndicator(),
-                  ),
-                ),
-              );
-            },
+          filteredLocalSnaps.when(
+            data: (data) => SliverList.builder(
+              itemCount: data.length,
+              itemBuilder: (context, index) => ManageSnapTile(
+                snap: data.elementAt(index),
+                position:
+                    determineTilePosition(index: index, length: data.length),
+              ),
+            ),
+            error: (_, __) =>
+                const SliverToBoxAdapter(child: SizedBox.shrink()),
+            loading: () => const SliverToBoxAdapter(
+              child: Center(
+                child: YaruCircularProgressIndicator(),
+              ),
+            ),
           ),
         ],
       ),
@@ -319,323 +320,6 @@ class _ActionButtons extends ConsumerWidget {
               overflow: TextOverflow.ellipsis,
             ),
           ),
-      ],
-    );
-  }
-}
-
-enum ManageTilePosition { first, middle, last, single }
-
-class _ManageSnapTile extends StatelessWidget {
-  const _ManageSnapTile({
-    required this.snap,
-    this.position = ManageTilePosition.middle,
-    this.showUpdateButton = false,
-  });
-
-  final Snap snap;
-  final ManageTilePosition position;
-  final bool showUpdateButton;
-
-  @override
-  Widget build(BuildContext context) {
-    final l10n = AppLocalizations.of(context);
-    final border = BorderSide(color: Theme.of(context).colorScheme.outline);
-    final dateTimeSinceUpdate = snap.installDate != null
-        ? DateTime.now().difference(snap.installDate!)
-        : null;
-    const radius = Radius.circular(8);
-
-    return DecoratedBox(
-      decoration: BoxDecoration(
-        borderRadius: switch (position) {
-          ManageTilePosition.first =>
-            const BorderRadius.only(topLeft: radius, topRight: radius),
-          ManageTilePosition.middle => BorderRadius.zero,
-          ManageTilePosition.last =>
-            const BorderRadius.only(bottomLeft: radius, bottomRight: radius),
-          ManageTilePosition.single => const BorderRadius.all(radius),
-        },
-        border: switch (position) {
-          ManageTilePosition.first => Border(
-              top: border,
-              left: border,
-              right: border,
-              bottom: border,
-            ),
-          ManageTilePosition.middle => Border(
-              left: border,
-              right: border,
-              bottom: border,
-            ),
-          ManageTilePosition.last => Border(
-              bottom: border,
-              left: border,
-              right: border,
-            ),
-          ManageTilePosition.single => Border.fromBorderSide(border),
-        },
-      ),
-      child: ListTile(
-        key: ValueKey(snap.id),
-        contentPadding: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
-        leading: Clickable(
-          onTap: () => StoreNavigator.pushSnap(context, name: snap.name),
-          child: AppIcon(iconUrl: snap.iconUrl, size: 40),
-        ),
-        title: Row(
-          children: [
-            Expanded(
-              flex: 2,
-              child: Align(
-                alignment: Alignment.centerLeft,
-                child: Clickable(
-                  onTap: () =>
-                      StoreNavigator.pushSnap(context, name: snap.name),
-                  child: Text(
-                    snap.titleOrName,
-                    maxLines: 1,
-                    overflow: TextOverflow.ellipsis,
-                  ),
-                ),
-              ),
-            ),
-            if (ResponsiveLayout.of(context).type !=
-                ResponsiveLayoutType.small) ...[
-              Expanded(
-                flex: 2,
-                child: dateTimeSinceUpdate != null
-                    ? Text(
-                        dateTimeSinceUpdate
-                            .managePageUpdateSinceDateTimeAgo(l10n),
-                        maxLines: 1,
-                        overflow: TextOverflow.ellipsis,
-                      )
-                    : const SizedBox(),
-              ),
-              Expanded(
-                child: snap.installedSize != null
-                    ? Text(
-                        context.formatByteSize(
-                          snap.installedSize!,
-                          precision: 0,
-                        ),
-                        maxLines: 1,
-                        overflow: TextOverflow.ellipsis,
-                      )
-                    : const SizedBox(),
-              ),
-            ],
-          ],
-        ),
-        subtitle: Column(
-          children: [
-            Row(
-              children: [
-                Text(snap.channel),
-                const SizedBox(width: 4),
-                Text(snap.version),
-              ],
-            ),
-            if (ResponsiveLayout.of(context).type == ResponsiveLayoutType.small)
-              Row(
-                children: [
-                  Expanded(
-                    child: dateTimeSinceUpdate != null
-                        ? Text(
-                            dateTimeSinceUpdate
-                                .managePageUpdateSinceDateTimeAgo(l10n),
-                            maxLines: 1,
-                            overflow: TextOverflow.ellipsis,
-                          )
-                        : const SizedBox(),
-                  ),
-                  Expanded(
-                    child: snap.installedSize != null
-                        ? Text(
-                            context.formatByteSize(
-                              snap.installedSize!,
-                              precision: 0,
-                            ),
-                            maxLines: 1,
-                            overflow: TextOverflow.ellipsis,
-                          )
-                        : const SizedBox(),
-                  ),
-                ],
-              ),
-          ],
-        ),
-        trailing: showUpdateButton
-            ? _ButtonBarForUpdate(snap)
-            : _ButtonBarForOpen(snap),
-      ),
-    );
-  }
-}
-
-ManageTilePosition determinePosition({
-  required int index,
-  required int length,
-}) {
-  if (length == 1) {
-    return ManageTilePosition.single;
-  }
-
-  if (index == length - 1) {
-    return ManageTilePosition.last;
-  }
-
-  if (index == 0) {
-    return ManageTilePosition.first;
-  } else {
-    return ManageTilePosition.middle;
-  }
-}
-
-class _ButtonBarForUpdate extends ConsumerWidget {
-  const _ButtonBarForUpdate(this.snap);
-
-  final Snap snap;
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final l10n = AppLocalizations.of(context);
-    final snapLauncher = ref.watch(launchProvider(snap));
-    final snapModel = ref.watch(snapModelProvider(snap.name));
-    final activeChangeId = snapModel.value?.activeChangeId;
-    final change = activeChangeId != null
-        ? ref.watch(activeChangeProvider(activeChangeId))
-        : null;
-
-    return Row(
-      mainAxisSize: MainAxisSize.min,
-      children: [
-        OutlinedButton(
-          onPressed: activeChangeId != null || !snapModel.hasValue
-              ? null
-              : ref.read(snapModelProvider(snap.name).notifier).refresh,
-          child: snapModel.value?.activeChangeId != null
-              ? Row(
-                  children: [
-                    SizedBox.square(
-                      dimension: kCircularProgressIndicatorHeight,
-                      child: YaruCircularProgressIndicator(
-                        value: change?.progress,
-                        strokeWidth: 2,
-                      ),
-                    ),
-                    if (change != null) ...[
-                      const SizedBox(width: 8),
-                      Text(
-                        change.localize(l10n) ?? '',
-                        maxLines: 1,
-                        overflow: TextOverflow.ellipsis,
-                      ),
-                    ],
-                  ],
-                )
-              : Row(
-                  mainAxisSize: MainAxisSize.min,
-                  children: [
-                    const Icon(YaruIcons.download),
-                    const SizedBox(width: 8),
-                    Text(
-                      l10n.snapActionUpdateLabel,
-                      maxLines: 1,
-                      overflow: TextOverflow.ellipsis,
-                    ),
-                  ],
-                ),
-        ),
-        const SizedBox(width: 8),
-        MenuAnchor(
-          menuChildren: [
-            Visibility(
-              maintainSize: true,
-              maintainAnimation: true,
-              maintainState: true,
-              visible: snapLauncher.isLaunchable,
-              child: MenuItemButton(
-                onPressed: snapLauncher.open,
-                child: Text(l10n.snapActionOpenLabel),
-              ),
-            ),
-            MenuItemButton(
-              onPressed: () =>
-                  StoreNavigator.pushSnap(context, name: snap.name),
-              child: Text(
-                l10n.managePageShowDetailsLabel,
-                style: Theme.of(context).textTheme.bodyMedium,
-              ),
-            ),
-          ],
-          builder: (context, controller, child) => YaruOptionButton(
-            onPressed: () {
-              if (controller.isOpen) {
-                controller.close();
-              } else {
-                controller.open();
-              }
-            },
-            child: const Icon(YaruIcons.view_more_horizontal),
-          ),
-        ),
-      ],
-    );
-  }
-}
-
-class _ButtonBarForOpen extends ConsumerWidget {
-  const _ButtonBarForOpen(this.snap);
-
-  final Snap snap;
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final snapLauncher = ref.watch(launchProvider(snap));
-    final l10n = AppLocalizations.of(context);
-
-    return Row(
-      mainAxisSize: MainAxisSize.min,
-      children: [
-        Visibility(
-          maintainSize: true,
-          maintainAnimation: true,
-          maintainState: true,
-          visible: snapLauncher.isLaunchable,
-          child: OutlinedButton(
-            onPressed: snapLauncher.open,
-            child: Text(
-              l10n.snapActionOpenLabel,
-            ),
-          ),
-        ),
-        const SizedBox(
-          width: 8,
-        ),
-        MenuAnchor(
-          menuChildren: [
-            MenuItemButton(
-              onPressed: () =>
-                  StoreNavigator.pushSnap(context, name: snap.name),
-              child: Text(
-                l10n.managePageShowDetailsLabel,
-                style: Theme.of(context).textTheme.bodyMedium,
-              ),
-            ),
-          ],
-          builder: (context, controller, child) => YaruOptionButton(
-            onPressed: () {
-              if (controller.isOpen) {
-                controller.close();
-              } else {
-                controller.open();
-              }
-            },
-            child: const Icon(YaruIcons.view_more_horizontal),
-          ),
-        ),
       ],
     );
   }
